requirements.txt
pillow
moviepy
librosa
numpy
matplotlib
ollama
skyfield
logging
numpy
asyncio
datetime
requests
gradio_client